G3: Fix: fixed the lint errors; added comments; unified the package naming; moved the generated pb files to pkg/service directory
