+++ Makefile
@@ -34,30 +34,30 @@ goose-install:
 	go install github.com/pressly/goose/v3/cmd/goose@latest
 
 generate_user_grpc: ## Generate grpc api
-	mkdir -p "pkg/user_v1"
+	mkdir -p "pkg/service/user_v1"
 	protoc --proto_path=api/user_v1 --proto_path vendor.protogen \
-  	--go_out=pkg/user_v1 --go_opt=paths=source_relative \
+  	--go_out=pkg/service/user_v1 --go_opt=paths=source_relative \
   	--plugin=protoc-gen-go=./bin/protoc-gen-go \
-  	--go-grpc_out=pkg/user_v1 --go-grpc_opt=paths=source_relative \
+  	--go-grpc_out=pkg/service/user_v1 --go-grpc_opt=paths=source_relative \
   	--plugin=protoc-gen-go-grpc=./bin/protoc-gen-go-grpc \
-  	--validate_out lang=go:pkg/user_v1 --validate_opt=paths=source_relative \
+  	--validate_out lang=go:pkg/service/user_v1 --validate_opt=paths=source_relative \
   	--plugin=protoc-gen-validate=bin/protoc-gen-validate \
-  	--grpc-gateway_out=pkg/user_v1 --grpc-gateway_opt=paths=source_relative \
+  	--grpc-gateway_out=pkg/service/user_v1 --grpc-gateway_opt=paths=source_relative \
   	--plugin=protoc-gen-grpc-gateway=bin/protoc-gen-go-gateway \
   	--openapiv2_out=allow_merge=true,merge_file_name=api:pkg/swagger \
   	--plugin=protoc-gen-openapiv2=bin/protoc-gen-openapiv2 \
   	api/user_v1/service.proto
 
 generate_user_grpc_win: ## Generate grpc api using windows protoc binaries
-	mkdir -p "pkg/user_v1"
+	mkdir -p "pkg/service/user_v1"
 	protoc --proto_path=api/user_v1 --proto_path vendor.protogen \
- 		--go_out=pkg/user_v1 --go_opt=paths=source_relative \
+ 		--go_out=pkg/service/user_v1 --go_opt=paths=source_relative \
 		--plugin=protoc-gen-go=./bin/protoc-gen-go.exe \
-		--go-grpc_out=pkg/user_v1 --go-grpc_opt=paths=source_relative \
+		--go-grpc_out=pkg/service/user_v1 --go-grpc_opt=paths=source_relative \
     	--plugin=protoc-gen-go-grpc=./bin/protoc-gen-go-grpc.exe \
-    	--validate_out lang=go:pkg/user_v1 --validate_opt=paths=source_relative \
+    	--validate_out lang=go:pkg/service/user_v1 --validate_opt=paths=source_relative \
     	--plugin=protoc-gen-validate=bin/protoc-gen-validate.exe \
-    	--grpc-gateway_out=pkg/user_v1 --grpc-gateway_opt=paths=source_relative \
+    	--grpc-gateway_out=pkg/service/user_v1 --grpc-gateway_opt=paths=source_relative \
     	--plugin=protoc-gen-grpc-gateway=bin/protoc-gen-grpc-gateway.exe \
     	--openapiv2_out=allow_merge=true,merge_file_name=api:pkg/swagger \
     	--plugin=protoc-gen-openapiv2=bin/protoc-gen-openapiv2.exe \
