+++ internal/api/user/service.go
@@ -2,16 +2,16 @@ package user
 
 import (
 	"context"
+	userV1 "github.com/Slintox/user-service/pkg/service/user_v1"
 
 	"google.golang.org/protobuf/types/known/emptypb"
 
 	converter "github.com/Slintox/user-service/internal/converter/user"
 	"github.com/Slintox/user-service/internal/service/user"
-	desc "github.com/Slintox/user-service/pkg/user_v1"
 )
 
 type Implementation struct {
-	desc.UnimplementedUserV1Server
+	userV1.UnimplementedUserV1Server
 
 	userService user.Service
 }
@@ -22,7 +22,7 @@ func NewImplementation(userService user.Service) *Implementation {
 	}
 }
 
-func (i *Implementation) Create(ctx context.Context, req *desc.CreateRequest) (*emptypb.Empty, error) {
+func (i *Implementation) Create(ctx context.Context, req *userV1.CreateRequest) (*emptypb.Empty, error) {
 	err := i.userService.Create(ctx, converter.ToCreateUserDesc(req))
 	if err != nil {
 		return nil, err
@@ -31,18 +31,18 @@ func (i *Implementation) Create(ctx context.Context, req *desc.CreateRequest) (*
 	return &emptypb.Empty{}, nil
 }
 
-func (i *Implementation) Get(ctx context.Context, req *desc.GetRequest) (*desc.GetResponse, error) {
+func (i *Implementation) Get(ctx context.Context, req *userV1.GetRequest) (*userV1.GetResponse, error) {
 	userView, err := i.userService.Get(ctx, req.GetUsername())
 	if err != nil {
 		return nil, err
 	}
 
-	return &desc.GetResponse{
+	return &userV1.GetResponse{
 		User: converter.FromUserDesc(userView),
 	}, nil
 }
 
-func (i *Implementation) Update(ctx context.Context, req *desc.UpdateRequest) (*emptypb.Empty, error) {
+func (i *Implementation) Update(ctx context.Context, req *userV1.UpdateRequest) (*emptypb.Empty, error) {
 	if req.UpdateData == nil {
 		return nil, errNoDataToUpdate
 	}
@@ -55,7 +55,7 @@ func (i *Implementation) Update(ctx context.Context, req *desc.UpdateRequest) (*
 	return &emptypb.Empty{}, nil
 }
 
-func (i *Implementation) Delete(ctx context.Context, req *desc.DeleteRequest) (*emptypb.Empty, error) {
+func (i *Implementation) Delete(ctx context.Context, req *userV1.DeleteRequest) (*emptypb.Empty, error) {
 	err := i.userService.Delete(ctx, req.GetUsername())
 	if err != nil {
 		return nil, err
