+++ internal/converter/user/converter.go
@@ -1,25 +1,25 @@
 package user
 
 import (
+	userV1 "github.com/Slintox/user-service/pkg/service/user_v1"
 	"google.golang.org/protobuf/types/known/timestamppb"
 
 	"github.com/Slintox/user-service/internal/model"
-	desc "github.com/Slintox/user-service/pkg/user_v1"
 )
 
 // FromUserDesc converts model.User -> grpc.User
-func FromUserDesc(user *model.User) *desc.User {
-	return &desc.User{
+func FromUserDesc(user *model.User) *userV1.User {
+	return &userV1.User{
 		Username:  user.Username,
 		Email:     user.Email,
-		RoleId:    desc.UserRoleID(user.RoleID),
+		RoleId:    userV1.UserRoleID(user.RoleID),
 		CreatedAt: timestamppb.New(user.CreatedAt),
 		UpdatedAt: timestamppb.New(user.UpdatedAt),
 	}
 }
 
 // ToCreateUserDesc converts grpc.CreateRequest -> model.User
-func ToCreateUserDesc(user *desc.CreateRequest) *model.CreateUser {
+func ToCreateUserDesc(user *userV1.CreateRequest) *model.CreateUser {
 	return &model.CreateUser{
 		Username:        user.GetUsername(),
 		Email:           user.GetEmail(),
@@ -30,7 +30,7 @@ func ToCreateUserDesc(user *desc.CreateRequest) *model.CreateUser {
 }
 
 // ToUpdateUserDesc converts grpc.UpdateUser -> model.UpdateUser
-func ToUpdateUserDesc(updateUserFields *desc.UpdateUserFields) *model.UpdateUser {
+func ToUpdateUserDesc(updateUserFields *userV1.UpdateUserFields) *model.UpdateUser {
 	updUser := &model.UpdateUser{
 		Username: updateUserFields.Username,
 		Email:    updateUserFields.Email,
